vm_patrol/requirements.txt
paramiko 
psycopg2-binary